refactor: keep QA helpers on private runtime seam
