+++ extensions/qa-lab/src/docker-runtime.ts
@@ -2,7 +2,7 @@ import {
   createQaDockerRuntime,
   type QaDockerFetchLike as FetchLike,
   type QaDockerRunCommand as RunCommand,
-} from "openclaw/plugin-sdk/qa-runner-shared-runtime";
+} from "openclaw/plugin-sdk/qa-runtime";
 
 export type { FetchLike, RunCommand };
