+++ extensions/qa-lab/src/live-transports/discord/cli.runtime.ts
@@ -1,4 +1,4 @@
-import { printLiveTransportQaArtifacts } from "openclaw/plugin-sdk/qa-runner-shared-runtime";
+import { printLiveTransportQaArtifacts } from "openclaw/plugin-sdk/qa-runtime";
 import type { LiveTransportQaCommandOptions } from "../shared/live-transport-cli.js";
 import { resolveLiveTransportQaRunOptions } from "../shared/live-transport-cli.runtime.js";
 import { runDiscordQaLive } from "./discord-live.runtime.js";
