+++ extensions/qa-lab/src/live-transports/slack/cli.runtime.ts
@@ -1,4 +1,4 @@
-import { printLiveTransportQaArtifacts } from "openclaw/plugin-sdk/qa-runner-shared-runtime";
+import { printLiveTransportQaArtifacts } from "openclaw/plugin-sdk/qa-runtime";
 import type { LiveTransportQaCommandOptions } from "../shared/live-transport-cli.js";
 import { resolveLiveTransportQaRunOptions } from "../shared/live-transport-cli.runtime.js";
 import { runSlackQaLive } from "./slack-live.runtime.js";
