+++ extensions/qa-lab/src/live-transports/telegram/cli.runtime.ts
@@ -1,4 +1,4 @@
-import { printLiveTransportQaArtifacts } from "openclaw/plugin-sdk/qa-runner-shared-runtime";
+import { printLiveTransportQaArtifacts } from "openclaw/plugin-sdk/qa-runtime";
 import type { LiveTransportQaCommandOptions } from "../shared/live-transport-cli.js";
 import { resolveLiveTransportQaRunOptions } from "../shared/live-transport-cli.runtime.js";
 import { listTelegramQaScenarioCatalog, runTelegramQaLive } from "./telegram-live.runtime.js";
