+++ extensions/qa-matrix/src/shared/live-transport-cli.runtime.test.ts
@@ -1,7 +1,7 @@
 import { mkdir, mkdtemp, readFile, rm } from "node:fs/promises";
 import os from "node:os";
 import path from "node:path";
-import { startLiveTransportQaOutputTee } from "openclaw/plugin-sdk/qa-runner-shared-runtime";
+import { startLiveTransportQaOutputTee } from "openclaw/plugin-sdk/qa-runtime";
 import { afterEach, describe, expect, it } from "vitest";
 
 const tmpDirs: string[] = [];
